Refactored web_parser for easier calls
