+++ src/Clean-Energy-PA-Site/EmailScheduler/tests/test_email_scheduler_base.py
@@ -2,8 +2,7 @@ from Website.tests.test_base import BaseTest
 from unittest import mock
 
 from web_parser.tests import zipsearch_test_data, ratesearch_test_data
-from web_parser.responses.ratesearch import offer_collection
-from web_parser.responses.zipsearch import distributor_collection
+from web_parser import offer_collection, distributor_collection
 
 
 # Create your tests here.
