+++ src/Clean-Energy-PA-Site/EmailScheduler/tests/test_views.py
@@ -3,14 +3,13 @@ from ..views import build_offer_path_less_than_rate, offer_search_less_than_rate
 from bs4 import BeautifulSoup
 from unittest import mock
 from web_parser.tests import zipsearch_test_data, ratesearch_test_data
-from web_parser.responses.ratesearch import offer_collection
-from web_parser.responses.zipsearch import distributor_collection
+from web_parser import offer_collection, distributor_collection
 
 
 class Email_Scheduler_View_Test(EmailSchedulerBaseTest):
-    @mock.patch("web_parser.papowerswitch_api.papowerswitch_api.get_offers")
+    @mock.patch("web_parser.papowerswitch_api.get_offers")
     def test_offers_results(self, mock_get_offers):
-        # Mock the distributors reponse
+        # Mock the distributors reponses
         mock_get_offers.return_value = offer_collection(
             ratesearch_test_data.expected_example
         )
