+++ src/web_parser/__init__.py
@@ -1,3 +1,7 @@
+from .papowerswitch_api import papowerswitch_api
+from .responses.ratesearch import offer, offer_collection, price_structure
+from .responses.zipsearch import distributor, distributor_collection
+
 import sys
 sys.path.append('.')
 sys.path.append('./tests')
